refactor(#508): extract shared runTest() logic to test-utils.ts
Consolidate ~860 lines of duplicated test runner code between test.ts
and test-worker.ts into a single TestUtils.runTest() function.

This also fixes three subtle inconsistencies that had diverged:
- test-worker.ts was missing update mode error→success conversion
- Validation order differed (NoWarnings vs Header ordering)
- test-worker.ts missing isTranspileOnly check on no-warnings

Both sequential (--jobs 1) and parallel execution modes verified.

Closes #508
